fix(font-ui): I guess we define the line-height as REM as well?
Not sure how much sense that make – on nzz.ch this size is being changed based on
the window width, so it would make more sense to just change the
font-size and then have the line-height be dependent on that? But
apparently this has not been designed like this.

We currently don't account for window width in our styles.